+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.66.2
 
+- keytrans: Detect new versions of account data in monitor responses and invoke search